LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-position-003-expected.txt


FAIL .target 1 assert_equals:
<div class="target" style="left: anchor(--a1 right)" data-offset-x="9"></div>
offsetLeft expected 9 but got 0
FAIL .target 2 assert_equals:
<div class="target" style="left: anchor(--a1 right)" data-offset-x="9"></div>
offsetLeft expected 9 but got 0
FAIL .target 3 assert_equals:
<div class="target" style="left: anchor(--a1 right)" data-offset-x="9"></div>
offsetLeft expected 9 but got 0
FAIL .target 4 assert_equals:
<div class="target" style="left: anchor(--a1 right)" data-offset-x="9"></div>
offsetLeft expected 9 but got 0
FAIL .target 5 assert_equals:
<div class="target" style="left: anchor(--a1 right)" data-offset-x="9"></div>
offsetLeft expected 9 but got 0
